Add joi-browser typing and move using by another typings
