+++ packages/core/src/declarations.d.ts
@@ -8,25 +8,9 @@ declare module 'electron-canvas-to-buffer' {
     export default _
 }
 
-declare module 'audiobuffer-to-wav' {
-    interface AudioBufferLike {
-        sampleRate: number
-        numberOfChannels: number
-        getChannelData: (channel: number) => Float32Array
-    }
-
-    const _: (buffer: AudioBufferLike, opts: { float32: boolean }) => ArrayBuffer
-    export = _
-}
-
-declare module 'arraybuffer-to-buffer' {
-    const _: (ab: ArrayBuffer) => Buffer
-    export = _
-}
-
-declare module 'keymirror' {
-    const _: (obj: {}) => { [key: string]: string }
-    export default _
+declare module 'joi-browser' {
+    import * as joi from 'joi'
+    export = joi
 }
 
 declare module 'node-timecodes' {
